salesflow-app/src/backend/requirements.txt
# ╔════════════════════════════════════════════════════════════════════════════╗
# ║  AURA OS - Python Backend Dependencies                                    ║
# ║  Install: pip install -r requirements.txt                                 ║
# ╚════════════════════════════════════════════════════════════════════════════╝

# ═══════════════════════════════════════════════════════════════════════════
# FASTAPI & SERVER
# ═══════════════════════════════════════════════════════════════════════════
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
gunicorn>=21.2.0  # Production WSGI Server

# ═══════════════════════════════════════════════════════════════════════════
# VALIDATION & SETTINGS
# ═══════════════════════════════════════════════════════════════════════════
pydantic>=2.5.0
pydantic-settings>=2.1.0

# ═══════════════════════════════════════════════════════════════════════════
# DATABASE (SUPABASE)
# ═══════════════════════════════════════════════════════════════════════════
supabase>=2.3.0

# ═══════════════════════════════════════════════════════════════════════════
# HTTP CLIENT
# ═══════════════════════════════════════════════════════════════════════════
httpx>=0.26.0
aiohttp>=3.9.0

# ═══════════════════════════════════════════════════════════════════════════
# AI / LLM
# ═══════════════════════════════════════════════════════════════════════════
anthropic>=0.18.0
openai>=1.12.0  # Fallback/Alternative

# ═══════════════════════════════════════════════════════════════════════════
# DOCUMENT PROCESSING (für Storybook Import)
# ═══════════════════════════════════════════════════════════════════════════
pymupdf>=1.23.0
python-docx>=1.1.0

# ═══════════════════════════════════════════════════════════════════════════
# PAYMENTS & MESSAGING
# ═══════════════════════════════════════════════════════════════════════════
stripe>=8.0.0            # Stripe Payments
twilio>=9.0.0            # SMS & WhatsApp

# ═══════════════════════════════════════════════════════════════════════════
# UTILITIES
# ═══════════════════════════════════════════════════════════════════════════
python-dotenv>=1.0.0
python-dateutil>=2.8.2

# ═══════════════════════════════════════════════════════════════════════════
# MONITORING & LOGGING (Production)
# ═══════════════════════════════════════════════════════════════════════════
sentry-sdk[fastapi]>=1.40.0  # Error Tracking
# json-logging>=1.5.0  # JSON Structured Logging (optional)

# ═══════════════════════════════════════════════════════════════════════════
# CACHING (Production)
# ═══════════════════════════════════════════════════════════════════════════
redis>=5.0.0  # Redis Client (optional)

# ═══════════════════════════════════════════════════════════════════════════
# TESTING (Dev only)
# ═══════════════════════════════════════════════════════════════════════════
pytest>=7.4.0
pytest-asyncio>=0.21.0
